requirements.txt
asgiref==3.4.1
Babel==2.9.1
Django==3.2.5
django-crispy-forms==1.12.0
django-phonenumber-field==5.2.0
phonenumbers==8.12.27
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
